pyproject.toml
[project]
name = "mcp-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.12.1",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "starlette>=0.52.1",
    "uvicorn>=0.41.0",
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "groq>=0.11.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]
